refactor: change name of atributes and methods to pass a better understandment, and reestructure folder tree to improove organization
