+++ lib/app/home/tabs/about/bloc/about_bloc.dart
@@ -5,15 +5,15 @@ import 'package:app_masterclass/app/home/tabs/about/models/dev_skills_model.dart
 import 'package:app_masterclass/app/home/tabs/about/models/favorite_tecnologies_model.dart';
 import 'package:app_masterclass/app/home/tabs/about/models/skill_model.dart';
 import 'package:app_masterclass/app/home/tabs/about/services/about_service/about_service.dart';
-import 'package:app_masterclass/app/home/tabs/about/services/skills_service/skills_service.dart';
 
 import '../models/about_model.dart';
+import '../services/technologies_and_skills_service/technologies_and_skills_service.dart';
 import 'about_events.dart';
 import 'about_state.dart';
 
 class AboutBloc {
   final IAboutService _aboutService;
-  final ISkillsService _skillsService;
+  final ITechnologiesAndSkillsService _technologiesAndSkillsService;
 
   final StreamController<AboutEvent> _inputAboutController =
       StreamController<AboutEvent>.broadcast();
@@ -23,7 +23,7 @@ class AboutBloc {
   Sink<AboutEvent> get inputAbout => _inputAboutController.sink;
   Stream<AboutState> get stream => _outputAboutController.stream;
 
-  AboutBloc(this._aboutService, this._skillsService) {
+  AboutBloc(this._aboutService, this._technologiesAndSkillsService) {
     _inputAboutController.stream.listen(_mapEventToState);
   }
 
@@ -33,8 +33,8 @@ class AboutBloc {
     if (event is LoadAboutEvent) {
       DevModel dev =
           await _aboutService.getGitHubUser(userName: event.userName);
-      DevSkillsModel devSkills =
-          await _skillsService.getDevSkills(url: dev.blog);
+      DevSkillsModel devSkills = await _technologiesAndSkillsService
+          .getDevTechnologiesAndSkills(url: dev.blog);
 
       List<FavoriteTecnologyModel> tecnologies = devSkills.favoritesTecnologies;
       List<SkillModel> skills = devSkills.skills;
