+++ lib/app/home/tabs/about/dependencies/dependency_injection.dart
@@ -1,16 +1,17 @@
 import 'package:app_masterclass/app/home/tabs/about/dependencies/about_injections.dart';
 import 'package:app_masterclass/app/home/tabs/about/services/about_service/dio_about_service.dart';
-import 'package:app_masterclass/app/home/tabs/about/services/skills_service/dio_skills_service.dart';
 import 'package:flutter/material.dart';
 
 import '../bloc/about_bloc.dart';
+import '../services/technologies_and_skills_service/dio_technologies_and_skills_service.dart';
 
 class DependencyInjection extends InheritedWidget {
   DependencyInjection({super.key, required super.child});
   //
 
   final AboutInjection aboutInjection = AboutInjection(
-      aboutBloc: AboutBloc(DioAboutService(), DioSkillsService()));
+      aboutBloc:
+          AboutBloc(DioAboutService(), DioTechnologiesAndSkillsService()));
 
   AboutBloc get aboutBloc => aboutInjection.aboutBloc;
