+++ lib/app/home/tabs/about/models/social_media_model.dart
@@ -1,4 +1,5 @@
 class SocialMedia {
+  //todo to implement on AboutModel
   final String name;
   final String logoUrl;
