--- lib/app/home/tabs/about/repositories/about_rpository.dart
@@ -1,1 +1,0 @@
-class AboutRepository {}
