--- lib/app/home/tabs/about/services/skills_service/skills_service.dart
@@ -1,5 +1,0 @@
-import 'package:app_masterclass/app/home/tabs/about/models/dev_skills_model.dart';
-
-abstract class ISkillsService {
-  Future<DevSkillsModel> getDevSkills({url});
-}
