+++ lib/app/home/tabs/about/services/technologies_and_skills_service/dio_technologies_and_skills_service.dart
@@ -1,12 +1,13 @@
 import 'package:app_masterclass/app/home/tabs/about/models/dev_skills_model.dart';
-import 'package:app_masterclass/app/home/tabs/about/services/skills_service/skills_service.dart';
 import 'package:app_masterclass/security/hasura_key.dart';
 
 import 'package:dio/dio.dart';
 
-class DioSkillsService extends ISkillsService {
+import 'technologies_and_skills_service.dart';
+
+class DioTechnologiesAndSkillsService extends ITechnologiesAndSkillsService {
   @override
-  Future<DevSkillsModel> getDevSkills({url}) async {
+  Future<DevSkillsModel> getDevTechnologiesAndSkills({url}) async {
     var dio = Dio();
     final response = await dio.get(
       url,
