+++ lib/app/home/tabs/about/services/technologies_and_skills_service/technologies_and_skills_service.dart
@@ -1,0 +1,5 @@
+import 'package:app_masterclass/app/home/tabs/about/models/dev_skills_model.dart';
+
+abstract class ITechnologiesAndSkillsService {
+  Future<DevSkillsModel> getDevTechnologiesAndSkills({url});
+}
