+++ lib/app/home/tabs/about/view/components/favorite_tecnologies/favorite_tecnology_widget.dart
