+++ lib/app/home/tabs/about/view/components/skills/skill_item_widget.dart
